Move "random" instantiators to random package
