+++ parcelabletest/src/main/java/com/backbase/test/instantiator/random/RandomEnumInstantiator.kt
@@ -1,5 +1,6 @@
-package com.backbase.test.instantiator
+package com.backbase.test.instantiator.random
 
+import com.backbase.test.instantiator.EnumInstantiator
 import java.util.Random
 
 /**
