+++ parcelabletest/src/main/java/com/backbase/test/instantiator/random/RandomLongInstantiator.kt
@@ -1,4 +1,4 @@
-package com.backbase.test.instantiator
+package com.backbase.test.instantiator.random
 
 import java.util.Random
